Disabling "Charging sounds" should prevent wireless sound as well
This is a fix for CYAN-4689 and also partial fix for CYAN-812. As noted in CYAN-812, separate section may not be a bad idea but the "Vibrate" checkbox already works with wireless charging so the sound settings should be allowed as well. There is no other (non complex) way for users to disable "wireless charging started" sound without modifying Android system.

Change-Id: I3311b94a7771863e49c1f83049600377dcf52d45